Use debootstrap to prepare the debian build root
Debian package build expect a correctly setup chroot, which includes
running the postinst script and doing a complete setup of all packages.
Current OBS mechanisms don't this automatically, at most only a
pre-configured subset of a packages are fully setup all the others are
simply unpacked. This causes build failures, switching to debootstrap
solves this as it knows how to setup a fully-functional debian chroot!

Signed-off-by: Héctor Orón Martínez <hector.oron@collabora.co.uk>