Double-check the debootstrap result
Debootstrap sometimes fails to setup packages, especially when the
dependencies aren't available. In this case it helpfully reports success
as it's exit code. Checking in the chroot if everything was left
unconfigured (and whether configuring that fails) detects these cases.

Also in such a case output the debootstrap log so it's easier to analyse
the problem.

Signed-off-by: Héctor Orón Martínez <hector.oron@collabora.co.uk>