Hack: Ignore negative dependencies
OBS normally treats the packages that the to be build package provides
as negative dependencies, iotw dependencies that should not be installed
for the build. This isn't needed for Debian (and would be a RC bug if a
package couldn't cope with this) and furthermore breaks debootstrap in
various cases.

Note this is a Hack, a proper patch would add a project configuration
item for this.

Signed-off-by: Héctor Orón Martínez <hector.oron@collabora.co.uk>