Ensure POSIX locale used in build
The build script sets LC_ALL to POSIX, but the Endless systemd package
then creates /etc/locale.conf (and the /etc/default/locale symlink) to
set LANG=en_US.utf8. Since the en_US.utf8 locale is not created in the
build root, we then get a bunch of warnings about setlocale failing.

Rather than try to get the locale defined, force the use of the POSIX
locale by setting it in the configuration.

Rebase 20180302 (T14726) - Move to build-recipe-debootstrap.