Create policy-rc.d in target to prevent daemon starting
If an optional /usr/sbin/policy-rc.d script exists, it will be run by
invoke-rc.d before deciding to start daemons. By exiting with code 101,
any actions are prevented from running. This effectively keeps daemons
from being started in the chroot and is needed when systemd is init
since the start-stop-daemon hack doesn't work in that case.

This is borrowed from the upstream sid debootstrap script, which took it
from debian-installer.

http://anonscm.debian.org/cgit/d-i/debootstrap.git/commit/?id=29cd6bee

https://phabricator.endlessm.com/T6209