Kill chroot processes with SIGTERM before SIGKILL
Killing all processes in the chroot with SIGKILL does not give them a
chance to clean up. Currently this is causing an issue with fakeroot
where faked is started and creates a semaphore for IPC. When faked is
killed with SIGKILL, it can't clean up the semaphore. Eventually, all
semaphores are exhausted and faked can't be started on a subsequent
build.

Call killchroot once with SIGTERM and wait 2 seconds before calling with
SIGKILL. Hopefully this will give processes enough time to cleanup
before they're forcefully killed.

https://phabricator.endlessm.com/T14181