Reorder packages before running scripts
The preinstall and install phases respect Order configuration, but the
intermediate runscripts phase didn't. Since this is likely to be the
first time programs are run inside the build root, the order in which
they're run is very important.

For example, on debian /etc/passwd and /etc/group don't exist until the
base-passwd postinst script is run. That needs to happen before the
base-files postinst script since it tries to chown files using names.
Without reordering, they'll be run in sort order with base-files first.

https://phabricator.endlessm.com/T14726