debootstrap: Move unmounts to cleanup hook
If the build is killed while the debootstrap root mounts are active,
then the umount calls would never be run. Move them to the
recipe_cleanup() hook, which is called by the main build script after
successful builds and for various cleanup routines including killing
running jobs. Most recipes do nothing during cleanup, but debootstrap is
unique in that it sets up a secondary build root.

https://phabricator.endlessm.com/T23364