debootstrap: Add abuild shadow password entry
Without this, calling `su` can fail during PAM authentication. The `su`
from shadow will ignore this error when called by root, but the `su`
from util-linux does not and the build fails.

The outer buildroot configured the abuild user correctly, but the copied
configuration for the inner debootstrap root did not.

https://phabricator.endlessm.com/T23585