build-recipe-dsc: pass the correct changes file to lintian
In the lintian checks section, we are passing (e.g.)
gem2deb_0.40_amd64.changes for a package where the release version was
set via dch and the name is actually gem2deb_0.40bem1_amd64.changes.

This causes lintian to fail and the build aborts.

Set OBS_DCH_RELEASE in the dch codepath so that the changes filename
is correctly constructed for lintian checks.

https://phabricator.endlessm.com/T25137